Title: 20대의 평균나이, Time: 1095ms, Memory: 0MB, Status: Time Limit Exceed - CodeTree

+++ 240802/20대의 평균나이/average-age-of-20.cpp
@@ -22,5 +22,5 @@ int main() {
 		cnt++;
 		sum += n;
 	}
-	cout << fixed << setprecision(2) << sum / cnt;
+	cout << fixed << setprecision(2) << (double)sum / cnt;
 }
